feat: add css styles
