--- src/components/BirdSubDescription/BirdSubDescription.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-export default function BirdSubDescription() {
-  return (
-    <>
-      <h3>Ястреб</h3>
-      <h4>Семейство</h4>
-      <audio controls>
-        <track kind="captions" />
-      </audio>
-    </>
-  )
-}
