+++ src/js/components/AnswerBlock/AnswerBlock.jsx
@@ -4,9 +4,9 @@ import BirdDescription from '../BirdDescription/BirdDescription'
 
 export default function AnswerBlock() {
   return (
-    <>
+    <div className="answers-block">
       <AnswersList />
       <BirdDescription />
-    </>
+    </div>
   );
 }
