+++ src/scss/_base.scss
@@ -1,0 +1,19 @@
+* {
+  margin: 0;
+  padding: 0;
+}
+
+html {
+  font-size: 10px;
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+  color: white;
+}
+
+ul {
+  list-style: none;
+}
+
+#root {
+  padding: 0 100px;
+  background-color: darkslategray;
+}
