+++ src/scss/_button.scss
@@ -1,0 +1,18 @@
+.button-next-level {
+  width: 100%;  
+  border: none;
+  border-radius: 10px;
+  color: white;
+  text-align: center;
+  font-size: 1.5rem;
+  padding: 10px;
+  margin-bottom: 30px;  
+  background-color: rgba(0, 0, 0, 0.3);
+
+  &_active {
+    background-color: #009d76;
+    cursor: pointer;
+  }
+}
+
+
